CHG: ability to remove duplicates for a subset of checksums [q12430]
git-svn-id: http://svn.resourcespace.com/svn/rs/trunk@19189 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ pages/tools/purge_duplicates.php
@@ -5,70 +5,134 @@ if('cli' != PHP_SAPI)
     exit('Access denied - Command line only!');
     }
 
-$webroot = dirname(dirname(__DIR__));
+$webroot = dirname(__DIR__, 2);
 include_once "{$webroot}/include/db.php";
-
 include_once "{$webroot}/include/log_functions.php";
+set_time_limit(0);
+logScript(sprintf('Started script - %s', __FILE__));
 
+$help_text = <<<'HELP'
+NAME
+    purge_duplicates - remove duplicate resources.
 
-// Script options (if required)
-$cli_short_options = '';
+SYNOPSIS
+    php /path/to/pages/tools/purge_duplicates.php [OPTIONS]
+
+DESCRIPTION
+    A tool to help administrators remove duplicate resources faster. If desired, duplicates can be fully deleted from 
+    the system.
+
+OPTIONS SUMMARY
+
+    -h, --help              Display this help text and exit.
+    --dry-run               Perform a trial run with no changes made.
+    --manage-method         Determines how the script will deal with the duplicates removal. Default: LIFO.
+                            Available options are:-
+                                * FIFO (First In, First Out) - will remove old resources and keep the last one found
+                                * LIFO (Last In, First Out) - will remove new duplicate resources and keep the first one found
+    --delete-permanently    Delete duplicate resources permanently. Default off - resources move to the configured resource_deletion_state
+    -c, --collection        Collection ID to limit the duplicates removal within a range (ie resources in the specified
+                            collection). You don't need to have all duplicates in the collection, just one of them will 
+                            be enough. Allowed multiple instances.
+
+EXAMPLES
+    # Testing script - always use dry-run to avoid silly mistakes (e.g typos)
+    php purge_duplicates.php --dry-run
+
+    # To remove all duplicates in the system (dropping the most recent duplicates)
+    php purge_duplicates.php
+    php purge_duplicates.php --manage-method=LIFO
+
+    # To remove all duplicates in the system (keeping the most recent duplicates)
+    php purge_duplicates.php --manage-method=FIFO
+
+    # To remove duplicates (keeping the most recent) in the system but only within a selected subset
+    php purge_duplicates.php --manage-method=FIFO --collection=50
+    php purge_duplicates.php --manage-method=FIFO -c=50 -c=100 -c=200
+
+HELP;
+
+// Script options @see https://www.php.net/manual/en/function.getopt.php
+$cli_short_options = 'hc:';
 $cli_long_options  = array(
+    'help',
+    'dry-run',
     'manage-method:',
     'delete-permanently',
+    'collection:'
 );
 
-/* managed-method option allows administrators to decide how the script will deal with the duplicates removal:
- - FIFO (First In, First Out) - will remove old resources and keep the last one found
- - LIFO (Last In, First Out) - will remove new duplicate resources and keep the first one found
-*/
+// Defaults
+$dry_run = false;
 $manage_method = 'LIFO';
-$order_by = 'r.ref ASC';
+$order_by = 'r.ref ASC'; # depends on the $manage_method value
 $delete_permanently = false;
+$collections = [];
 
 foreach(getopt($cli_short_options, $cli_long_options) as $option_name => $option_value)
     {
-    if($option_name == 'manage-method' && !is_array($option_value) && in_array($option_value, array('FIFO', 'LIFO')))
+    if(in_array($option_name, ['h', 'help']))
+        {
+        ob_clean();
+        echo $help_text;
+        exit(0);
+        }
+    else if(in_array($option_name, ['dry-run', 'delete-permanently']))
+        {
+        logScript("Script running with '{$option_name}' option enabled!");
+        $option_name = str_replace("-", "_", $option_name);
+        $$option_name = true;
+        }
+    else if($option_name == 'manage-method' && is_string($option_value) && in_array($option_value, array('FIFO', 'LIFO')))
         {
         $manage_method = $option_value;
-
-        if($option_value == 'FIFO')
+        if($manage_method == 'FIFO')
             {
             $order_by = 'r.ref DESC';
             }
-
-        continue;
         }
-
-    if($option_name == 'delete-permanently')
+    else if(in_array($option_name, ['c', 'collection']))
         {
-        $delete_permanently = true;
+        $collections = array_values(array_filter(is_array($option_value) ? $option_value : [$option_value], 'is_int_loose'));
+        logScript("Script running within subset(s) range. Collections received: " . implode(', ', $collections));
         }
     }
+logScript("Script running with the '{$manage_method}' method!");
 
-logScript("Purging duplicates using '{$manage_method}' method!");
 
-$duplicates = sql_query("
+$collections_limit_sql = empty($collections) 
+        ? ''
+        : "-- IF file checksum is within a specified range
+       AND file_checksum IN (
+                SELECT DISTINCT r.file_checksum
+                  FROM resource AS r
+            INNER JOIN collection_resource AS cr ON r.ref = cr.resource
+                 WHERE cr.collection IN (" . ps_param_insert(count($collections)) . ")
+                   AND (file_checksum IS NOT NULL AND file_checksum <> '')
+              GROUP BY r.file_checksum
+           )
+";
+$duplicates = ps_query("
         SELECT r.ref, r.file_extension, r.file_checksum
           FROM resource AS r
          WHERE length(r.file_extension) > 0
            AND (r.file_checksum IS NOT NULL AND r.file_checksum <> '')
+           {$collections_limit_sql}
+           -- IF duplicate file checksum (ie. more than one resource has it)
            AND file_checksum IN (
                     SELECT file_checksum
-                      FROM (
-                                SELECT file_checksum
-                                  FROM resource
-                                 WHERE (file_checksum IS NOT NULL AND file_checksum <> '')
-                              GROUP BY file_checksum HAVING count(file_checksum) > 1
-                            ) AS rfc2
+                      FROM resource
+                     WHERE (file_checksum IS NOT NULL AND file_checksum <> '')
+                  GROUP BY file_checksum HAVING count(file_checksum) > 1
                )
-      ORDER BY {$order_by} 
-");
+      ORDER BY {$order_by}
+    ",
+    ps_param_fill($collections, 'i')
+);
 logScript('Found #' . count($duplicates) . ' duplicates');
 logScript("");
 
-$saved_duplicates = array(); # Key is the resource ID and value is the file checksum
-
+$saved_duplicates = []; # Key is the resource ID and value is the file checksum
 foreach($duplicates as $duplicate)
     {
     logScript("Processing resource #{$duplicate['ref']} having checksum '{$duplicate['file_checksum']}'");
@@ -84,14 +148,19 @@ foreach($duplicates as $duplicate)
     if($delete_permanently)
         {
         logScript("Deleting permanently resource #{$duplicate['ref']}");
-
-        unset($resource_deletion_state);
-        delete_resource($duplicate['ref']);
+        if(!$dry_run)
+            {
+            unset($resource_deletion_state);
+            delete_resource($duplicate['ref']);
+            }
         }
     else
         {
         logScript("Moving duplicate resource #{$duplicate['ref']} to archive state #{$resource_deletion_state}");
-        update_archive_status($duplicate['ref'], $resource_deletion_state);
+        if(!$dry_run)
+            {
+            update_archive_status($duplicate['ref'], $resource_deletion_state);
+            }
         }
     }
